fixed plugin bot ensure
